+++ R/Versions.R
@@ -786,7 +786,7 @@ getPackageBinaryURL <- function(packageName, version = NULL, repos = "https://cl
 
 getPackageFileExt <- function(platform = NA, type = c("binary", "source")) {
     
-    type <- match.arg(type)
+    type <- RstoxData::match_arg_informative(type)
     platform <- getPlatform(platform)
     
     if (platform == "windows") {