+++ inst/tinytest/test-split_18.R
@@ -1,3 +1,3 @@
 # Run the test project:
 projectPaths <- system.file("test",  "splitNASC_18.zip", package = "RstoxFramework")
-expect_true(compareProjectToStoredOutputFiles(projectPaths))
+expect_true(compareProjectToStoredOutputFiles(projectPaths, ignore = c("MinLayerDepth", "MaxLayerDepth")))