+++ inst/tinytest/test-tobis_20.R
@@ -1,3 +1,3 @@
 # Run the test project:
 projectPaths <- system.file("test",  "tobis_20.zip", package = "RstoxFramework")
-expect_true(compareProjectToStoredOutputFiles(projectPaths))
+expect_true(compareProjectToStoredOutputFiles(projectPaths, ignore = c("MinLayerDepth", "MaxLayerDepth", "Layer")))